added validators: Session and Appointments
