+++ src/app/controllers/SessionController.js
@@ -1,5 +1,4 @@
 import jwt from 'jsonwebtoken';
-import * as Yup from 'yup';
 
 import authConfig from '../../config/auth';
 import User from '../models/User';
@@ -7,17 +6,6 @@ import File from '../models/File';
 
 class SessionController {
   async store(req, resp) {
-    const schema = Yup.object().shape({
-      email: Yup.string()
-        .email()
-        .required(),
-      password: Yup.string().required(),
-    });
-
-    if (!(await schema.isValid(req.body))) {
-      return resp.status(400).json({ error: 'Validation fails' });
-    }
-
     const { email, password } = req.body;
 
     const user = await User.findOne({
