+++ src/app/controllers/UserController.js
@@ -1,4 +1,3 @@
-import * as Yup from 'yup';
 import User from '../models/User';
 import File from '../models/File';
