+++ src/app/validators/AppointmentStore.js
@@ -1,0 +1,18 @@
+import * as Yup from 'yup';
+
+export default async (req, resp, next) => {
+  try {
+    const schema = Yup.object().shape({
+      provider_id: Yup.number().required(),
+      date: Yup.date().required(),
+    });
+
+    await schema.validate(req.body, { abortEarly: false });
+
+    return next();
+  } catch (err) {
+    return resp
+      .status(400)
+      .json({ error: 'Validation fails', messages: err.inner });
+  }
+};
